PP-6381 Merge charge 'model' into transaction service
The charge model actually behaved like a service. The methods in it
belong in transaction.service alongside other methods that deal with
operations on transactions. Move the methods here and delete the charge
model.
